struggle

+++ transformer_engine/common/triton/mhc.py
@@ -239,12 +239,11 @@ def _mhc_projection_bwd_fused(
     )  # (BLOCK_SIZE_M, BLOCK_SIZE_K)
     grad_x_ptrs = grad_x_ptr + offs_m[:, None] * stride_grad_xm + offs_k[None, :] * stride_grad_xk
     grad_x = grad_x.to(x.dtype)
-    if FUSE_GRAD_X_ACC:
+    if FUSE_GRAD_X_ACC: # If fused gradient accumulation is enabled, the buffer is always fp32
         grad_x_acc = tl.load(
             grad_x_ptrs, mask=mask_m[:, None] & mask_k[None, :], other=0.0
         )
-        grad_x = grad_x.to(tl.float32) + grad_x_acc.to(tl.float32)
-        grad_x = grad_x.to(x.dtype)
+        grad_x = grad_x.to(tl.float32) + grad_x_acc
     tl.store(grad_x_ptrs, grad_x, mask=mask_m[:, None] & mask_k[None, :])
 
 def projection_config_bwd_norm_weight():
@@ -1081,17 +1080,12 @@ def _mhc_aggregate_bwd(
 
     grad_x_ptrs = grad_x_ptr + offs_m[:, None] * stride_grad_xm + offs_cn[None, :] * stride_grad_xCn
 
-    if FUSE_GRAD_X_ACC:
+    if FUSE_GRAD_X_ACC: # If fused gradient accumulation is enabled, the buffer is always fp32
         grad_x_acc = tl.load(
             grad_x_ptrs, mask=mask_m[:, None] & mask_cn[None, :], other=0.0
         )
-        grad_x = grad_x.to(tl.float32) + grad_x_acc.to(tl.float32)
-        grad_x = grad_x.to(x.dtype)
-    tl.store(
-        grad_x_ptrs,
-        grad_x,
-        mask=mask_m[:, None] & mask_cn[None, :],
-    )
+        grad_x = grad_x.to(tl.float32) + grad_x_acc
+    tl.store(grad_x_ptrs, grad_x, mask=mask_m[:, None] & mask_cn[None, :],)
 
 
 def expand_combine_config():
